+++ Implementation/microrestjs-framework/lib/platform/Microrest.js
@@ -10,10 +10,10 @@
  * @module
  */
 
-var winston = require('winston');
 var checkTypes = require('check-types');
 
-var credentialsGenerator = require('./utils/CredentialsGenerator');
+var credentialsGenerator = require('../helpers/security/CredentialsGenerator');
+var client = require('./Client');
 
 /**
  * Port where the server is listening.
@@ -74,15 +74,15 @@ module.exports.getInstance = function getInstance() {
  */
 function Microrest() {
     //Initializes the internal state
-    this.configuration = require('./ConfigurationLoader').loadConfiguration();
-    _configureLogger(this.configuration.logger);
+    this.configuration = require('../helpers/loaders/ConfigurationLoader').loadConfiguration();
+    require('../helpers/logging/LoggerManager').configure(this.configuration.logger);
 
     port = this.configuration.server.port;
     defaultDirectoryLocation = this.configuration.directory.location;
 
-    //Generates and publishes the platform credentials asynchronously.
-    this.arePublishedCredentials = false;
-    credentialsGenerator.generateCredentials(this._getPublishCredentialsFunction());
+    //Generates the platform credentials asynchronously.
+    this.platformCredentials = null;
+    credentialsGenerator.generateCredentials(this._getStoreCredentialsFunction());
 
     this.server = require('./Server').getInstance();
     this.serviceManager = require('./ServiceManager').getInstance();
@@ -94,33 +94,51 @@ function Microrest() {
 }
 
 /**
- * Gets the function to publish the generated credentials in:
+ * Gets the function to store the generated credentials as the credentials of the platform.
+ *
+ * @private
+ * @function
+ */
+Microrest.prototype._getStoreCredentialsFunction = function _getStoreCredentialsFunction() {
+    var _this = this;
+
+    return function _storeCredentials(error, credentials) {
+        if (checkTypes.assigned(error)) {
+            throw new Error('The platform credentials could not be generated, because ' + error.message);
+        }
+
+        if (checkTypes.not.object(credentials) || checkTypes.emptyObject(credentials) ||
+            checkTypes.not.string(credentials.key) || checkTypes.not.unemptyString(credentials.key) ||
+            checkTypes.not.string(credentials.certificate) || checkTypes.not.unemptyString(credentials.certificate)) {
+            throw new Error('The generated platform credentials are not correct, because either the key or the certificate is not present.');
+        }
+
+        _this.platformCredentials = credentials;
+    };
+};
+
+/**
+ * Publishes the credentials of the platform in:
  *    - the Server to listen SSL requests.
  *    - the Client to make SSL requests.
  *    - the RunnableServices to know their certificate.
  *
  * @private
  * @function
+ * @returns {Boolean} - true, if the credentials have been published; false, otherwise.
  */
-Microrest.prototype._getPublishCredentialsFunction = function _getPublishCredentialsFunction() {
-    var _this = this;
-    return function _publishCredentials(error, credentials) {
-            if (checkTypes.assigned(error)) {
-                throw new Error('The platform credentials could not be generated, because ' + error.message);
-            }
+Microrest.prototype._publishCredentials = function _publishCredentials() {
+    if (checkTypes.not.object(this.platformCredentials) || checkTypes.emptyObject(this.platformCredentials) ||
+        checkTypes.not.string(this.platformCredentials.key) || checkTypes.not.unemptyString(this.platformCredentials.key) ||
+        checkTypes.not.string(this.platformCredentials.certificate) || checkTypes.not.unemptyString(this.platformCredentials.certificate)) {
+        return false;
+    }
 
-            if (checkTypes.not.object(credentials) || checkTypes.emptyObject(credentials) ||
-                checkTypes.not.string(credentials.key) || checkTypes.not.unemptyString(credentials.key) ||
-                checkTypes.not.string(credentials.certificate) || checkTypes.not.unemptyString(credentials.certificate)) {
-                throw new Error('The generated platform credentials are not correct, because either the key or the certificate is not present.');
-            }
+    this.server.addPlatformCredentials(this.platformCredentials);
+    this.serviceManager.addCertificateToServices(this.platformCredentials.certificate);
+    client.addPlatformCredentials(this.platformCredentials);
 
-            _this.server.addCredentials(credentials);
-            _this.serviceManager.addCertificateToServices(credentials.certificate);
-            require('./Client').addCredentials(credentials);
-
-            _this.arePublishedCredentials = true;
-        };
+    return true;
 };
 
 /**
@@ -130,7 +148,8 @@ Microrest.prototype._getPublishCredentialsFunction = function _getPublishCredent
  * @function
  */
 Microrest.prototype.run = function run() {
-    if (this.arePublishedCredentials !== true) {
+    var areCredentialsPublished = this._publishCredentials();
+    if (areCredentialsPublished !== true) {
         setTimeout(run.bind(this), 1000);
         return;
     }
@@ -153,32 +172,13 @@ Microrest.prototype.shutdown = function shutdown() {
         this.serviceManager.shutdown();
     }
 
+    client.cleanPlatformCredentials();
+
     this.configuration = null;
+    this.platformCredentials = null;
     this.server = null;
     this.serviceManager = null;
 
     port = 0;
     defaultDirectoryLocation = '';
 };
-
-/**
- * Configures the Winston logger used by the Microrestjs Framework
- *
- * @private
- * @function
- */
-function _configureLogger(loggerConfiguration) {
-    if (loggerConfiguration.enable === false) {
-        winston.loggers.options.transports = [
-            new (winston.transports.Console)({
-                level: 'none'
-            })
-        ];
-    } else {
-        winston.loggers.options.transports = [
-            new (winston.transports.Console)({
-                level: loggerConfiguration.level
-            })
-        ];
-    }
-}